+++ src/olmo_core/internal/experiment.py
@@ -443,7 +443,7 @@ def build_config(
 def launch(config: ExperimentConfig):
     log.info(config)
     assert config.launch is not None
-    config.launch.launch(follow=True)
+    config.launch.launch(follow=True, launch_timeout=5 * 60)
 
 
 def launch_prep(config: ExperimentConfig):