+++ src/scripts/train/OLMo3/OLMo3-7B.py
@@ -6,7 +6,7 @@ from olmo_core.distributed.parallel import DataParallelType
 from olmo_core.float8 import Float8Config
 from olmo_core.internal.common import CLUSTER_TO_GPU_TYPE
 from olmo_core.internal.experiment import CommonComponents, build_config, main
-from olmo_core.nn.attention import SlidingWindowAttentionConfig
+from olmo_core.nn.attention import AttentionBackendName
 from olmo_core.nn.transformer import TransformerConfig
 from olmo_core.optim import CosWithWarmup, OptimGroupOverride, SkipStepAdamWConfig
 from olmo_core.train import Duration, TrainerConfig
@@ -22,13 +22,10 @@ GLOBAL_BATCH_SIZE = 4 * 1024 * 1024
 
 
 def build_model_config(common: CommonComponents) -> TransformerConfig:
-    config = TransformerConfig.olmo2_7B(vocab_size=common.tokenizer.padded_vocab_size())
-    config.block.attention.sliding_window = SlidingWindowAttentionConfig(
-        force_full_attention_on_first_layer=False,
-        force_full_attention_on_last_layer=True,
-        pattern=[4096, 4096, 4096, -1],
+    config = TransformerConfig.olmo3_7B(
+        vocab_size=common.tokenizer.padded_vocab_size(),
+        attn_backend=AttentionBackendName.flash_2,
     )
-    config.block.attention.use_flash = True
     return config